# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/eventCounts": {
      "get": {
        "description": "Queries the custom-events cache for the given tenant and time range, then groups events by their `name` field and counts occurrences. Returns an array of `{ value, count }` pairs sorted by count descending — useful for a ranked event-type breakdown.",
        "operationId": "EventsController_getEventCounts",
        "parameters": [
          {
            "name": "start",
            "required": true,
            "in": "query",
            "description": "Start of the time range (ISO 8601 string or epoch milliseconds).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end",
            "required": true,
            "in": "query",
            "description": "End of the time range (ISO 8601 string or epoch milliseconds).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantId",
            "required": true,
            "in": "query",
            "description": "Tenant identifier used to scope the cache query.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "/eventCounts",
        "tags": [
          "v1"
        ]
      }
    },
    "/widgetsByName": {
      "get": {
        "description": "Filters the custom-events cache to events matching the provided `eventName` query parameter, then groups the results by `data.widgetName` and counts occurrences. Returns an array of `{ value, count }` pairs sorted by count descending — useful for per-widget interaction breakdowns.",
        "operationId": "EventsController_getEventCountsByName",
        "parameters": [
          {
            "name": "eventName",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start",
            "required": true,
            "in": "query",
            "description": "Start of the time range (ISO 8601 string or epoch milliseconds).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end",
            "required": true,
            "in": "query",
            "description": "End of the time range (ISO 8601 string or epoch milliseconds).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantId",
            "required": true,
            "in": "query",
            "description": "Tenant identifier used to scope the cache query.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "/widgetsByName",
        "tags": [
          "v1"
        ]
      }
    },
    "/events": {
      "get": {
        "description": "Returns all custom events from the cache within the given time range for the specified tenant. Each event contains `name`, `date`, and `data`. Set `withId=true` to also include `identifyId` and `sessionId` on each event object.",
        "operationId": "EventsController_getEvents",
        "parameters": [
          {
            "name": "start",
            "required": true,
            "in": "query",
            "description": "Start of the time range (ISO 8601 string or epoch milliseconds).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end",
            "required": true,
            "in": "query",
            "description": "End of the time range (ISO 8601 string or epoch milliseconds).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantId",
            "required": true,
            "in": "query",
            "description": "Tenant identifier used to scope the cache query.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "withId",
            "required": false,
            "in": "query",
            "description": "When true, includes identifyId and sessionId in each event object. Defaults to false.",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "/events",
        "tags": [
          "v1"
        ]
      }
    },
...
    "/v2/events": {
      "get": {
        "description": "Supports optional server-side filtering via filtrex (https://github.com/cshaa/filtrex), field projection, and data sub-field projection. All string values in filter expressions must be double-quoted.",
        "operationId": "EventsControllerV2_getEventsV2",
        "parameters": [
          {
            "name": "start",
            "required": true,
            "in": "query",
            "description": "Start of the time range (ISO 8601 string or epoch milliseconds).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end",
            "required": true,
            "in": "query",
            "description": "End of the time range (ISO 8601 string or epoch milliseconds).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantId",
            "required": true,
            "in": "query",
            "description": "Tenant identifier used to scope the cache query.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "filtrex filter expression applied to mapped event objects. Strings must be double-quoted.\n\n**Examples**\n- `name == \"buttonClick\"` — button click events only\n- `name != \"pageView\"` — exclude page view events\n- `name in (\"buttonClick\", \"formSubmit\")` — clicks and form submissions\n- `identifyId == \"user-001\"` — events from a specific user",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fields",
            "required": false,
            "in": "query",
            "description": "Comma-separated top-level event fields to return. Valid values: name, date, data, identifyId, sessionId. Omit to return all fields.\n\n**Examples**\n- `name,date,identifyId` — event name, timestamp, and user\n- `name,data` — event name and raw payload only\n- `name,date,sessionId` — event name, timestamp, and session context",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dataFields",
            "required": false,
            "in": "query",
            "description": "Comma-separated dot-notation paths to project from event.data. Omit to return the full data object.\n\n**Examples**\n- `widgetName` — widget name only\n- `widgetName,widgetId` — widget name and identifier\n- `attributes.size,attributes.color` — specific nested attribute paths",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "description": "Field and optional direction to sort results by, in the format `field` or `field:asc` / `field:desc`. Direction defaults to `asc` when omitted. Valid field values match those listed in `fields`.\n\n**Examples**\n- `date:desc` — most recent events first\n- `date:asc` — oldest events first\n- `name:asc` — alphabetical by event name\n- `identifyId:asc` — alphabetical by user",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Filtered and projected event objects.",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "name": "buttonClick",
                      "date": "2024-01-15T10:30:00.000Z",
                      "data": {
                        "widgetName": "SaveButton",
                        "widgetId": "save-btn"
                      },
                      "identifyId": "user-001",
                      "sessionId": "sess-abc123"
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "basic": []
          }
        ],
        "summary": "v2/events",
        "tags": [
          "v2"
        ]
      }
    },
...
  },
  "info": {
    "title": "Gainsight Sync Microservice API",
    "description": "A NestJS microservice that periodically pulls Gainsight PX telemetry (custom events, session events, page views, active-user metrics) from configured tenant domains, caches the data in memory per tenant using a date-sorted array with binary-search range queries, and exposes REST endpoints consumed by dashboards such as Grafana.\n\nAll data endpoints require `Authorization: Basic <base64(user:password)>` and a `tenantId` query parameter. Public endpoints: `/health`, `/lastRun`, `/llms.txt`, `/llms-full.txt`.\n\nInteractive docs: https://gainsight.eu-latest.cumulocity.com/apps/gainsight-c8y-openapi/index.html",
    "version": "0.2.2",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://gainsight.eu-latest.cumulocity.com/service/gainsight-sync-ms"
    }
  ],
  "components": {
    "securitySchemes": {
      "basic": {
        "type": "http",
        "scheme": "basic"
      }
    },
...
  }
}